+++ __init__.py
@@ -1,5 +1,0 @@
-"""Initialize PyPI Package"""
-
-# pylint: disable=invalid-name
-
-__all__ = ["main"]