+++ __main__.py
@@ -1,8 +1,0 @@
-"""The main entry point for the project."""
-
-# pylint: disable=invalid-name, import-error
-
-from main import timer
-
-if __name__ == "__main__":
-    timer()