+++ src/moms_canning_timer/__main__.py
@@ -1,0 +1,8 @@
+"""Run with ``python -m moms_canning_timer``."""
+
+# Copyright (C) 2017-2026 willtheorangeguy
+
+from .app import main
+
+if __name__ == "__main__":
+    main()